+++ src/theme/main.ts
@@ -1,0 +1,8 @@
+/**
+ * Main
+ * A main content element that expands to fill the full viewport height.
+ * ---
+ */
+export default {
+  base: 'min-h-[calc(100vh-var(--b24ui-header-height))]'
+}