+++ CHANGELOG.md
@@ -25,6 +25,15 @@
 * chore(deps): update @nuxt/test-utils + vitest to v4
 * fix(components): add `fixed` prop to prevent responsive text size reduction
 * fix(types): improve `DotPathKeys` accuracy and `GetItemKeys` performance
+* feat(dictionary): add menu icon
+* docs(CSS Variables): improve
+* feat(Theme): new component
+* feat(Footer): new component
+* feat(FooterColumns): new component
+* feat(Header): new component
+* feat(Main): new component
+* feat(Drawer): new component
+
 
 ## [2.3.0](https://github.com/bitrix24/b24ui/compare/v2.2.1...v2.3.0) (2026-02-12)